Optimize GP operations: replace_child, math.log, complexity cache, bugfixes
- replace_child() + child_index property on SymbolicNode (tree.py),
  11 call sites refactored in gpoperator.py — eliminates list+tuple allocations
- np.log → math.log (3× in gpoperator.py) — restores lru_cache to 100% hit
- Remove redundant update_constants in optimizer objectives (gpoperator.py)
- ExpressionSet._complexity_cache with invalidation in update_constants
- Defer current_individuals list comp into is_multi_output_ branch (base.py)
- Remove redundant result.astype(np.float32) in _sigmoid (node.py)
- find_top_n boundary fix when n >= len(fitnesses) (population.py)

8 inserts → 2 inserts shuffled; 5 bugfixes; 72 fewer lines.
All 174 tests pass. HOF output is bitwise identical to baseline.

+++ symbolearn/node.py
@@ -1283,7 +1283,7 @@ def _sigmoid(x: np.ndarray) -> np.ndarray:
         result[x <= EXP_LOWER_BOUND] = 0.0
         result[x >= EXP_UPPER_BOUND] = 1.0
 
-        return result.astype(np.float32)
+        return result
 
 
 def _softplus(x: np.ndarray) -> np.ndarray:
